refactor: split channel routes into per-channel handler files
